Revert "[expo-notifications][iOS] Swift conversion 4: Categories (#33976)"
This reverts commit f93af371430c8fd1d60383ae12e814964bf3056f.
