+++ packages/expo-notifications/ios/EXNotifications/Notifications/Categories/EXNotificationCategoriesModule.h
@@ -1,0 +1,25 @@
+// Copyright 2018-present 650 Industries. All rights reserved.
+
+#import <ExpoModulesCore/EXExportedModule.h>
+#import <EXNotifications/EXNotificationsDelegate.h>
+
+@interface EXNotificationCategoriesModule : EXExportedModule <EXNotificationsDelegate>
+
+- (void)getNotificationCategoriesAsyncWithResolver:(EXPromiseResolveBlock)resolve reject:(EXPromiseRejectBlock)reject;
+- (void)setNotificationCategoryWithCategoryId:(NSString *)categoryId
+                                      actions:(NSArray *)actions
+                                      options:(NSDictionary *)options
+                                      resolve:(EXPromiseResolveBlock)resolve 
+                                       reject:(EXPromiseRejectBlock)reject;
+- (void)deleteNotificationCategoryWithCategoryId:(NSString *)categoryId
+                                         resolve:(EXPromiseResolveBlock)resolve 
+                                          reject:(EXPromiseRejectBlock)reject;
+- (NSMutableDictionary *)serializeCategory:(UNNotificationCategory *)category;
+- (UNNotificationCategory *)createCategoryWithId:(NSString*)categoryId
+                                         actions:(NSArray *)actions
+                                         options:(NSDictionary *)options;
+- (NSMutableDictionary *)serializeCategoryOptions:(UNNotificationCategory *)category;
+- (NSMutableArray *)serializeActions:(NSArray<UNNotificationAction *>*)actions;
+- (NSMutableDictionary *)serializeActionOptions:(NSUInteger)options;
+
+@end
