+++ packages/expo-notifications/ios/EXNotifications/Notifications/Scheduling/SchedulerModule.swift
@@ -33,6 +33,81 @@ let dateComponentsMatchMap: [String: Calendar.Component] = [
   "weekdayOrdinal": .weekdayOrdinal
 ]
 
+struct CalendarTriggerRecord: Record {
+  @Field
+  var year: Int?
+  @Field
+  var month: Int?
+  @Field
+  var day: Int?
+  @Field
+  var hour: Int?
+  @Field
+  var minute: Int?
+  @Field
+  var second: Int?
+  @Field
+  var weekday: Int?
+  @Field
+  var weekOfMonth: Int?
+  @Field
+  var weekOfYear: Int?
+  @Field
+  var weekdayOrdinal: Int?
+  @Field
+  var timezone: String?
+  @Field
+  var repeats: Bool?
+}
+
+struct TimeIntervalTriggerRecord: Record {
+  @Field
+  var seconds: TimeInterval
+  @Field
+  var repeats: Bool
+}
+
+struct DateTriggerRecord: Record {
+  @Field
+  var timestamp: TimeInterval
+}
+
+struct DailyTriggerRecord: Record {
+  @Field
+  var hour: Int
+  @Field
+  var minute: Int
+}
+
+struct WeeklyTriggerRecord: Record {
+  @Field
+  var weekday: Int
+  @Field
+  var hour: Int
+  @Field
+  var minute: Int
+}
+
+struct MonthlyTriggerRecord: Record {
+  @Field
+  var day: Int
+  @Field
+  var hour: Int
+  @Field
+  var minute: Int
+}
+
+struct YearlyTriggerRecord: Record {
+  @Field
+  var month: Int
+  @Field
+  var day: Int
+  @Field
+  var hour: Int
+  @Field
+  var minute: Int
+}
+
 public class SchedulerModule: Module {
   public func definition() -> ModuleDefinition {
     Name("ExpoNotificationScheduler")
